Refactor code and improve CLI interface of `ctk info` vs. `ctk cfr`
