+++ cratedb_toolkit/cli.py
@@ -7,11 +7,11 @@ from .adapter.rockset.cli import cli as rockset_cli
 from .cfr.cli import cli as cfr_cli
 from .cluster.cli import cli as cloud_cli
 from .cmd.tail.cli import cli as tail_cli
+from .info.cli import cli as info_cli
 from .io.cli import cli as io_cli
 from .job.cli import cli_list_jobs
 from .query.cli import cli as query_cli
 from .shell.cli import cli as shell_cli
-from .wtf.cli import cli as wtf_cli
 
 
 @click.group(cls=ClickAliasedGroup)  # type: ignore[arg-type]
@@ -23,6 +23,7 @@ def cli(ctx: click.Context, verbose: bool, debug: bool):
     return boot_click(ctx, verbose, debug)
 
 
+cli.add_command(info_cli, name="info")
 cli.add_command(cfr_cli, name="cfr")
 cli.add_command(cloud_cli, name="cluster")
 cli.add_command(io_cli, name="load")
@@ -30,5 +31,4 @@ cli.add_command(query_cli, name="query")
 cli.add_command(rockset_cli, name="rockset")
 cli.add_command(shell_cli, name="shell")
 cli.add_command(tail_cli, name="tail")
-cli.add_command(wtf_cli, name="wtf")
 cli.add_command(cli_list_jobs)
