+++ cratedb_toolkit/info/http.py
@@ -7,10 +7,10 @@ from functools import lru_cache
 import typing_extensions as t
 from fastapi import Depends, FastAPI, HTTPException
 
+from cratedb_toolkit.info.core import InfoContainer
+from cratedb_toolkit.info.util import get_baseinfo
 from cratedb_toolkit.util import DatabaseAdapter
 from cratedb_toolkit.util.service import start_service
-from cratedb_toolkit.wtf.core import InfoContainer
-from cratedb_toolkit.wtf.util import get_baseinfo
 
 logger = logging.getLogger(__name__)
 
@@ -38,4 +38,4 @@ def info(category: str, adapter: t.Annotated[DatabaseAdapter, Depends(database_a
 
 
 def start(listen_address: t.Union[str, None] = None, reload: bool = False):  # pragma: no cover
-    start_service(app="cratedb_toolkit.wtf.http:app", listen_address=listen_address, reload=reload)
+    start_service(app="cratedb_toolkit.info.http:app", listen_address=listen_address, reload=reload)
