+++ tests/conftest.py
@@ -21,8 +21,8 @@ RESET_TABLES = [
     '"doc"."jobinfo"',
     '"ext"."clusterinfo"',
     '"ext"."jobinfo"',
-    f'"{TESTDRIVE_EXT_SCHEMA}"."qc_statement_log"',
-    f'"{TESTDRIVE_EXT_SCHEMA}"."qc_last_execution"',
+    f'"{TESTDRIVE_EXT_SCHEMA}"."jobstats_statements"',
+    f'"{TESTDRIVE_EXT_SCHEMA}"."jobstats_last"',
     f'"{TESTDRIVE_EXT_SCHEMA}"."retention_policy"',
     f'"{TESTDRIVE_DATA_SCHEMA}"."raw_metrics"',
     f'"{TESTDRIVE_DATA_SCHEMA}"."sensor_readings"',
